requirements.txt
pandas
openai
python-dotenv
ipykernel
shapely
requests
streamlit
folium
-e .